requirements.txt
opencv-python
tensorflow
flask
flask-socketio
numpy
pillow
google-generativeai
python-dotenv
mediapipe
matplotlib
seaborn
python-engineio
python-socketio
eventlet
